VER3-162 move validator to shared
